Don’t try to make filtering general-purpose
